Animation observers must opt in to get notified when the animator is destroyed.
With this change, animation observers will not care about getting NotifyAborted notifications when the animator is destroyed, unless they explicitly opt into this behavior. If an observer does decide to override RequiresNotificationWhenAnimatorDestroyed to return true, it must also ensure that it removes itself as an observer before it is destroyed. There is a strongly worded comment to this effect over RequiresNotificationWhenAnimatorDestroyed.

BUG=111362
TEST=Verify fix for issue 111362

Review URL: http://codereview.chromium.org/9169087

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119271 0039d316-1c4b-4281-b951-d872f2087c98