aura compact mode continuous alt-tab should land on a window instead of showing the background.
This fix the problem when we alt-tab too fast, offscreen windows are not hidden yet, therefore we may/may not get a visiblity change event to update the current window. Also, should set the current window before animation starts, so that when animation end and we started hiding windows, we don't accidentally hide the current window.

BUG=111211
TEST=tested on alex that background doesn't show if alt-tab is continuously down.


Review URL: http://codereview.chromium.org/9169039

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@119418 0039d316-1c4b-4281-b951-d872f2087c98