Store the UploadData identifier in NavigationEntryImpl.
This will ultimately be used to help save content that was retrieved from a POST.

A few points:
  - The identifier can be recovered from the TabRestoreService
  - The identifier is not persisted to disk for either the session or the tab service.
  
The identifier is redundant with data in the "opaque" state_ string, but we decided to go this route. See http://groups.google.com/a/chromium.org/group/chromium-dev/browse_thread/thread/9f521b7c18850b3b/ for a discussion of alternative approaches.

BUG=55551
TEST=None

Review URL: http://codereview.chromium.org/9317009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120369 0039d316-1c4b-4281-b951-d872f2087c98