Move Ownership of IMUI to HistoryService.
This change moves the InMemoryURLIndex (IMUI) from being owned by the InMemoryHistoryBackend to being owned by the HistoryService. The only dependency the IMUI has on the HistoryBackend is the need to access the history database for the one-time need to build the IMUI from scratch. Divorcing the IMUI from the HistoryBackend also pointed out that updates to the IMUI index related to URL visits and history deletes should be handled by notification, and this was implemented in this CL.

NOTE: Portions previously reviewed as: http://codereview.chromium.org/9030031/. That change involved both 1) moving the ownership of the IMUI from the InMemoryHistoryBackend, and 2) performing the caching file operations on the FILE thread. I've split this up into two CLs with this CL dealing with the move of ownership.

BUG=83659
TEST=Updated unit tests.
Review URL: https://chromiumcodereview.appspot.com/9316109

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120707 0039d316-1c4b-4281-b951-d872f2087c98