Early-return in RenderViewImpl::DidFlushPaint() if there is no WebView.
CL 9328016 exposed a bug with seemingly-unused stats code assuming that the RenderView's WebView must be valid when DidFlushPaint() is called, which may not be true if RenderWidget::OnSwapBuffersComplete is notified while the widget is closing down.

We'll follow up by verifying the stats code is unnecessary and removing it (see bug 112942).

BUG=112921,111639,112942


Review URL: http://codereview.chromium.org/9349004

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120715 0039d316-1c4b-4281-b951-d872f2087c98