Added daemon process NPAPI.
This provides a framework that we can implement per-platform to allow the
Me2Me daemon to installed, started and stopped by the web-app.

I've gone with a polling-based API for now, as it simplifies some of
the logic. We can revisit this if necessary to make it more sophisticated.

BUG=None
TEST=None until we have some implementations.


Review URL: http://codereview.chromium.org/9316078

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@120723 0039d316-1c4b-4281-b951-d872f2087c98