Fix crashing timing bug where panel animates after its closed.
There is a timing gap between when a panel is closed and when its removed from
panel strip. This results in a crash when an event happens, say a shuffle, that
results in an animation on the closed panel.

Fixed this by removing the panel from panel strip on Close rather than in
destructor. Also moved other cleanup code from destructor to Close.

BUG=111120
TEST=Steps in bug report.

Review URL: https://chromiumcodereview.appspot.com/9317025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@121054 0039d316-1c4b-4281-b951-d872f2087c98