Intoroducing Motoya font for ChromeOS
This is a retry of r120926, which was reverted.  The error was due to
duplicate entry in the translation resource file, which only happens
on 'Official' build.

We will introduce Motoya fonts (MotoyaG04Gothic, MotoyaG04Mincho)
as the default Japanese font for ChromeOS.  This CL is for preparing
resource to define default font and Skia to check the font family
is compatible with their counterparts of MS Windows fonts.

Note that GTK font specification in the resource does not
handle multiple fonts (e.g. "MotoyaG04Gothic 10, IPAPGothic 10")
and I left it as is (see crosbug.com/17382 for details).
Fontconfig should cover the compatibility check
(IPAPGothic => MotoyaG04Gothic) and display the proper font.

BUG=chromium-os:16690
TEST=build, run and see Japanese fonts are properly displayed in both web contents and Chrome UI.


Review URL: http://codereview.chromium.org/9365012

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@121170 0039d316-1c4b-4281-b951-d872f2087c98