aura: Display black background when locking from wrench.
When the screen is locked via the wrench menu instead of by
holding the power button, we skip the slow-close animation
and don't show the black background layer.  This formerly
caused the compositor's blue clear color to be visible and
now results in a neat hall-of-mirrors effect due to previous
frames not being cleared.

This change makes us show the background if it isn't already
visible when we start the fast-close animation after getting
notice that we're locking.

BUG=113526
TEST=added, also manual


Review URL: http://codereview.chromium.org/9385030

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@121825 0039d316-1c4b-4281-b951-d872f2087c98