Chrome on Android: Remove speech input files from android build.
Speech input is not currently implemented on Android (no support for FLAC,
etc...), so we want to remove the related speech files from the android gypi
definitions.

Also, ifdefs the single usage point of the SpeechInputManager to only be
constructed if ENABLE_INPUT_SPEECH is enabled in gyp.

BUG=113487
TEST=

Review URL: http://codereview.chromium.org/9347057

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@122903 0039d316-1c4b-4281-b951-d872f2087c98