Windows Breakpad: Stop doing Breakpad initialisation on a thread
The Breakpad setup was being done on a thread because reading metadata
chrome.dll was slow, but the code has been changed to read the
metadata from chrome.exe which should be faster.

I need to remove the use of threading so that I can change the
Breakpad enablement checks to happen only in the browser process and
be inherited reliably by subprocesses.

Also, this will potentially make any crashes on startup more reliably
reported by Breakpad.

BUG=111265
TEST=none


Review URL: http://codereview.chromium.org/9423016

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@123149 0039d316-1c4b-4281-b951-d872f2087c98