sync: Delete ScopedDirLookup and DirectoryManager
These classes were simple wrappers that didn't actually do a whole lot.
They added an unnecessary layer of indirection and required us to write
several lines of boilerplate every time we wanted to access the
directory.

The DirectoryManager did manage some locks, but it is believe that these
are no longer necessary.  The directory management functions (ie. open,
close, save changes, etc.) are only invoked from the syncer thread and
all other operations require a transaction.  Directory operations should
be safe even without these locks.

The DirectoryManager was responsible for taking the input directory path
argument, appending a constant filename, and passing the result on to
the Directory.  The constant value has been moved to a static member of
syncable::Directory and it is now the responsibility of the caller of
Directory::Open() to assemble the full file path.

Deleting the DirectoryManager left the Cryptographer without a home.  It
was moved into the syncable::Directory for now.  This allows us to
perform more thorough runtime verification of the locking assertions
around this class.

These changes had some reprecussions for the unit test base classes.
The TestDirectorySetterUpper underwent significant modifications.
Hopefully these will make it easier to read and reduce duplication of
test boilerplate code.

This change also wound up affecting the SyncerCommandTest class.  The
template-based conditional instantiation of a Directory or a
MockDirectory was making it hard to use the TestDirectorySetterUpper, so
the class hierarchy was refactored to allow for greater code reuse.

BUG=115560
TEST=


Review URL: http://codereview.chromium.org/9442031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@123897 0039d316-1c4b-4281-b951-d872f2087c98