Don't retry failed attempts to poll sync server
This commit doesn't change much of our intended behaviour, though it
does affect the details of how we achieve it.

A failed poll used to cause us to schedule a retry job.  However,
because this job was considered to be a POLL job, and we don't run POLL
jobs when we're in backoff, it would get rejected by the scheduler.  The
backoff timer never got reset after that, so this would have happened
only once per poll attempt.

This change makes it so we never backoff in response to a failed poll.
This might have an impact in some really strange edge cases (ie. a poll
recently failed then we receive an IP address change notification) but
those changes should be an improvement and most common use cases will be
unaffected.  Users will likely never notice the effects of this change.

Our tests, however, do care about these sorts of details.  This commit
fixes some existing tests and adds one new one.

BUG=112954
TEST=SyncSchedulerTest in sync_unit_test


Review URL: http://codereview.chromium.org/9372051

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@124034 0039d316-1c4b-4281-b951-d872f2087c98