Make the lifetime delegate optional.
PepperPluginRegistry appears to be the only class to implement a PluginDelegate::ModuleLifetime.

The Browser Plugin is accessed through a webkit::ppapi::PluginModule which takes in a lifetime delegate.

The Browser Plugin does not make use of PepperPluginRegistry (as suggested by brettw), and so
we checking for a non-NULL delegate before accessing it in PluginModule.

BUG=none
TEST=manually


Review URL: http://codereview.chromium.org/9618045

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@125734 0039d316-1c4b-4281-b951-d872f2087c98