Fix some problems with extension keywords and cases where the AutocompleteController asked for something other than ALL_MATCHES.  Also some misc. cleanup.
The first change in the file is clearly important; the others are less clear.  Mainly they amount to running the ApplyDefaultSuggestion call all the time, which should be harmless according to Matt, and returning the previous result set (with modified scores) for the "minimal changes" case even for SYNCHRONOUS_ONLY, which should also be safe.  (We exclude the BEST_MATCH case because if we're knocking the old matches' scores down below the main match, then clearly none of them will become the best match.)

BUG=none
TEST=Install omnibox extension, don't enable in incognito.  Open an incognito window, type the extension's keyword and some input string (e.g. "omnix foo" for the "omnix" sample omnibox extension) into the address bar.  Note no results from extension.  Switch tabs without clearing address bar.  Switch back, focus address bar, hit enter.  You should search, rather than just doing nothing.
Review URL: https://chromiumcodereview.appspot.com/9693028

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@126300 0039d316-1c4b-4281-b951-d872f2087c98