Add strace_inputs.py to strace a test executable and detect its dependencies.
It detects if all the files in a directories were read and put the dependency on
the directory entry instead of listing each individual files.

Detects all the temporary files that are created in the checkout directory and
that will need to be cleaned up to mark http://crbug.com/116251 fixed.

R=rogerta@chromium.org
BUG=98640
TEST=


Review URL: http://codereview.chromium.org/9689025

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@126395 0039d316-1c4b-4281-b951-d872f2087c98