Change the pixel tests reference image revision mechanism.
Originally we put the revision in the file name.  However, this blocks the flakiness dashboard to obtain the images.  So now we keep a seperate revision file in the format of testname_revision.ref, whereas the ref image is always testname.png no matter what's the revision.

BUG=
TEST=GPU bots, pixel tests
R=jbates
Review URL: https://chromiumcodereview.appspot.com/9692052

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@126457 0039d316-1c4b-4281-b951-d872f2087c98