Use XDamageSubtract to fetch the damaged region from the X server.
Previously the damaged area was accumulated based on received XDamage events, and the entire desktop then "repaired" with an XDamageSubtract() call, leading to a race-condition between damage notifications and the repair.

This CL registers only for non-empty damage notifications, and atomically fetches and clears the damage region using XDamageSubtract(), to ensure that no damage is missed.

BUG=116283,117485
TEST=Change the resolution of a Chromoting Me2Me host between two sizes and verify that it is always displayed correctly at the client.


Review URL: http://codereview.chromium.org/9700084

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@127257 0039d316-1c4b-4281-b951-d872f2087c98