Ensure the previously active window gets the focus after a Chrome window is closed on OSX.
This is a rework of the previous fix in r120593 to handle non-browser windows as well.

New logic relies on Panel browser windows returning false to -canBecomeKeyWindow in most cases.
App could be left without a key window as a result, so I also added logic to ensure that there is a key window in this case.

BUG=112038
TEST=Manually open/close tabbed windows, panels, Task Manager. Check focus after closing windows.

Review URL: https://chromiumcodereview.appspot.com/9417033

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@127517 0039d316-1c4b-4281-b951-d872f2087c98